thesis/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
networkx==3.3
matplotlib==3.9.2
scipy==1.15.2
typing_extensions==4.11.0
psutil==7.0.0
py-cpuinfo==9.0.0